Vanilla icu4c 76.1

+++ source/test/cintltst/udatatst.c
@@ -46,6 +46,7 @@
 #include "ucol_imp.h"
 #include "ucol_swp.h"
 #include "ucnv_bld.h"
+#include "udataswp.h"
 #include "sprpimpl.h"
 #include "rbbidata.h"
 
@@ -1147,7 +1148,7 @@ static void TestICUDataName(void)
                 typeChar = 'l';
                 break;
           default:
-                log_err("Expected 1 or 0 for U_IS_BIG_ENDIAN, got %d!\n", (int)U_IS_BIG_ENDIAN);
+                log_err("Expected 1 or 0 for U_IS_BIG_ENDIAN, got %d!\n", U_IS_BIG_ENDIAN);
                 /* return; */
           }
           break;
